+++ modules/alloy/src/main/ts/ephox/alloy/api/Main.ts
@@ -16,7 +16,7 @@ import * as AlloyParts from '../parts/AlloyParts';
 import * as PartType from '../parts/PartType';
 import * as Bubble from '../positioning/layout/Bubble';
 import * as Layout from '../positioning/layout/Layout';
-import * as LayoutInside from '../positioning/layout/LayoutInside';
+import * as LayoutInset from '../positioning/layout/LayoutInset';
 import * as LayoutTypes from '../positioning/layout/LayoutTypes';
 import * as MaxHeight from '../positioning/layout/MaxHeight';
 import * as MaxWidth from '../positioning/layout/MaxWidth';
@@ -246,7 +246,7 @@ export {
 
   // layout
   Layout,
-  LayoutInside,
+  LayoutInset,
   LayoutTypes,
   Bubble,
   MaxHeight,