+++ modules/tinymce/src/themes/silver/main/ts/backstage/Anchors.ts
@@ -5,7 +5,7 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { AlloyComponent, Bubble, HotspotAnchorSpec, Layout, LayoutInside, MaxHeight, NodeAnchorSpec, SelectionAnchorSpec } from '@ephox/alloy';
+import { AlloyComponent, Bubble, HotspotAnchorSpec, Layout, LayoutInset, MaxHeight, NodeAnchorSpec, SelectionAnchorSpec } from '@ephox/alloy';
 import { Optional } from '@ephox/katamari';
 import { SimSelection, SugarElement, SugarShadowDom } from '@ephox/sugar';
 
@@ -36,8 +36,8 @@ const getInlineDialogAnchor = (contentAreaElement: () => SugarElement, lazyAncho
     node: Optional.from(contentAreaElement()),
     bubble: Bubble.nu(bubbleSize, bubbleSize, bubbleAlignments),
     layouts: {
-      onRtl: () => [ LayoutInside.northeast ],
-      onLtr: () => [ LayoutInside.northwest ]
+      onRtl: () => [ LayoutInset.northeast ],
+      onLtr: () => [ LayoutInset.northwest ]
     },
     overrides
   });
@@ -62,8 +62,8 @@ const getBannerAnchor = (contentAreaElement: () => SugarElement, lazyAnchorbar: 
     root: SugarShadowDom.getContentContainer(contentAreaElement()),
     node: Optional.from(contentAreaElement()),
     layouts: {
-      onRtl: () => [ LayoutInside.north ],
-      onLtr: () => [ LayoutInside.north ]
+      onRtl: () => [ LayoutInset.north ],
+      onLtr: () => [ LayoutInset.north ]
     }
   });